user profile

+++ lib/models/image_format_def.dart
@@ -21,7 +21,7 @@ class FormatDef {
   String mime;
   String name;
   String path;
-  int size;
+  double size;
   int width;
   int height;
